requirements.txt
rasa
torch
pydub
nltk
pymorphy2
openai-whisper
pytorch_lightning
transformers
ffmpeg
fpdf